libinst: factor out cmdline.[ch]
