+++ libinst/gpib.h
@@ -1,19 +1,12 @@
 #ifndef _INST_GPIB_H
 #define _INST_GPIB_H 1
 
-#include <getopt.h>
-
 typedef struct gpib_device *gd_t;
 
 /* Callback function to interpret results of serial poll.
  * Return value: -1=!ready, 0=success, >0=fatal error
  */
 typedef int (*spollfun_t)(gd_t gd, unsigned char status_byte, char *msg);
-
-/* Look up a device's default address in /etc/gpib-utils.conf
- * Result is allocated in static storage that may be overwritten on next call.
- */
-char *gpib_default_addr(char *name);
 
 /* Initialize/finalize a device.  If sf is non-NULL, a serial poll is 
  * run after every I/O and the resulting status byte is passed to the sf
@@ -64,41 +57,6 @@ void gpib_trg(gd_t gd);
 int gpib_rsp(gd_t gd, unsigned char *status);
 
 void gpib_abort(gd_t gd); /* VXI only */
-
-typedef struct {
-    char *sopt;
-    char *lopt;
-    char *desc;
-} opt_desc_t;
-
-void usage(opt_desc_t *tab);
-
-#define GPIB_UTILS_CONF { \
-    "$GPIB_UTILS_CONF", \
-    "~/.gpib-utils.conf", \
-    "/etc/gpib-utils.conf", \
-   NULL, \
-}
-
-#define OPTIONS_COMMON "a:n:v"
-#define OPTIONS_COMMON_LONG \
-    {"address",         required_argument, 0, 'a'}, \
-    {"name",            no_argument,       0, 'n'}, \
-    {"verbose",         no_argument,       0, 'v'}
-
-#define OPTIONS_COMMON_DESC \
-    {"a", "address",   "set instrument address" }, \
-    {"n", "name",      "set instrument name" }, \
-    {"v", "verbose",   "show protocol on stderr" }
-
-#define OPTIONS_COMMON_SWITCH \
-    case 'a': \
-    case 'n': \
-    case 'v':
-
-gd_t gpib_init_args(int argc, char *argv[], const char *opts, 
-                    struct option *longopts, char *name, 
-                    spollfun_t sf, unsigned long retry, int *opt_error);
 
 #endif /* !INST_GPIB_H */
